Move all `ConfigurationBuilder` functionality to `Configuration`
