+++ src/statue/cli/config/config_general.py
@@ -3,7 +3,7 @@ import click
 
 from statue.cli.common_flags import config_path_option
 from statue.cli.config.config_cli import config_cli
-from statue.config.configuration_builder import ConfigurationBuilder
+from statue.config.configuration import Configuration
 from statue.runner import RunnerMode
 
 
@@ -17,8 +17,8 @@ from statue.runner import RunnerMode
 def set_mode_cli(mode, config):
     """Choose which runner mode will be used by default."""
     if config is None:
-        config = ConfigurationBuilder.configuration_path()
-    configuration = ConfigurationBuilder.from_file(config)
+        config = Configuration.configuration_path()
+    configuration = Configuration.from_file(config)
     configuration.default_mode = RunnerMode[mode]
     configuration.to_toml(config)
     click.echo("Mode was successfully set!")
@@ -30,8 +30,8 @@ def set_mode_cli(mode, config):
 def set_history_size_cli(size, config):
     """Choose which runner mode will be used by default."""
     if config is None:
-        config = ConfigurationBuilder.configuration_path()
-    configuration = ConfigurationBuilder.from_file(config)
+        config = Configuration.configuration_path()
+    configuration = Configuration.from_file(config)
     configuration.cache.history_size = size
     configuration.to_toml(config)
     click.echo("History size was successfully set!")
