+++ tests/configuration/configuration/test_configuration_build_from_dict.py
@@ -6,7 +6,6 @@ from pytest_cases import parametrize
 
 from statue.config.commands_repository import CommandsRepository
 from statue.config.configuration import Configuration
-from statue.config.configuration_builder import ConfigurationBuilder
 from statue.config.contexts_repository import ContextsRepository
 from statue.config.sources_repository import SourcesRepository
 from statue.constants import (
@@ -24,9 +23,7 @@ from statue.runner import RunnerMode
 
 def test_configuration_builder_from_empty_config(tmp_path):
     cache_dir = tmp_path / ".statue"
-    configuration = ConfigurationBuilder.from_dict(
-        cache_dir=cache_dir, statue_config_dict={}
-    )
+    configuration = Configuration.from_dict(cache_dir=cache_dir, statue_config_dict={})
 
     assert isinstance(configuration, Configuration)
     assert len(configuration.contexts_repository) == 0
@@ -38,9 +35,9 @@ def test_configuration_builder_from_empty_config(tmp_path):
 
 
 @parametrize(argnames="mode", argvalues=[RunnerMode.SYNC, RunnerMode.ASYNC])
-def test_configuration_builder_from_dict_mode_upper_case(mode, tmp_path):
+def test_configuration_from_dict_mode_upper_case(mode, tmp_path):
     cache_dir = tmp_path / ".statue"
-    configuration = ConfigurationBuilder.from_dict(
+    configuration = Configuration.from_dict(
         cache_dir=cache_dir, statue_config_dict={GENERAL: {MODE: mode.name.upper()}}
     )
 
@@ -53,9 +50,9 @@ def test_configuration_builder_from_dict_mode_upper_case(mode, tmp_path):
 
 
 @parametrize(argnames="mode", argvalues=[RunnerMode.SYNC, RunnerMode.ASYNC])
-def test_configuration_builder_from_dict_mode_lower_case(mode, tmp_path):
+def test_configuration_from_dict_mode_lower_case(mode, tmp_path):
     cache_dir = tmp_path / ".statue"
-    configuration = ConfigurationBuilder.from_dict(
+    configuration = Configuration.from_dict(
         cache_dir=cache_dir, statue_config_dict={GENERAL: {MODE: mode.name.lower()}}
     )
 
@@ -67,10 +64,10 @@ def test_configuration_builder_from_dict_mode_lower_case(mode, tmp_path):
     assert configuration.default_mode == mode
 
 
-def test_configuration_builder_from_dict_history_size(tmp_path):
+def test_configuration_from_dict_history_size(tmp_path):
     cache_dir = tmp_path / ".statue"
     size = random.randint(1, 100)
-    configuration = ConfigurationBuilder.from_dict(
+    configuration = Configuration.from_dict(
         cache_dir=cache_dir, statue_config_dict={GENERAL: {HISTORY_SIZE: size}}
     )
 
@@ -82,14 +79,14 @@ def test_configuration_builder_from_dict_history_size(tmp_path):
     assert configuration.default_mode == RunnerMode.SYNC
 
 
-def test_configuration_builder_from_dict_update_contexts(tmp_path):
+def test_configuration_from_dict_update_contexts(tmp_path):
     contexts_config = mock.Mock()
     cache_dir = tmp_path / ".statue"
 
     with mock.patch.object(
         ContextsRepository, "update_from_config"
     ) as contexts_update_mock:
-        configuration = ConfigurationBuilder.from_dict(
+        configuration = Configuration.from_dict(
             cache_dir=cache_dir, statue_config_dict={CONTEXTS: contexts_config}
         )
         contexts_update_mock.assert_called_once_with(contexts_config)
@@ -102,14 +99,14 @@ def test_configuration_builder_from_dict_update_contexts(tmp_path):
     assert configuration.default_mode == RunnerMode.SYNC
 
 
-def test_configuration_builder_from_dict_update_commands(tmp_path):
+def test_configuration_from_dict_update_commands(tmp_path):
     commands_config = mock.Mock()
     cache_dir = tmp_path / ".statue"
 
     with mock.patch.object(
         CommandsRepository, "update_from_config"
     ) as commands_update_mock:
-        configuration = ConfigurationBuilder.from_dict(
+        configuration = Configuration.from_dict(
             cache_dir=cache_dir, statue_config_dict={COMMANDS: commands_config}
         )
         commands_update_mock.assert_called_once_with(commands_config)
@@ -122,14 +119,14 @@ def test_configuration_builder_from_dict_update_commands(tmp_path):
     assert configuration.default_mode == RunnerMode.SYNC
 
 
-def test_configuration_builder_from_dict_update_sources(tmp_path):
+def test_configuration_from_dict_update_sources(tmp_path):
     sources_config = mock.Mock()
     cache_dir = tmp_path / ".statue"
 
     with mock.patch.object(
         SourcesRepository, "update_from_config"
     ) as sources_update_mock:
-        configuration = ConfigurationBuilder.from_dict(
+        configuration = Configuration.from_dict(
             cache_dir=cache_dir, statue_config_dict={SOURCES: sources_config}
         )
         sources_update_mock.assert_called_once_with(
@@ -144,11 +141,11 @@ def test_configuration_builder_from_dict_update_sources(tmp_path):
     assert configuration.default_mode == RunnerMode.SYNC
 
 
-def test_configuration_builder_from_dict_fail_update_mode(tmp_path):
+def test_configuration_from_dict_fail_update_mode(tmp_path):
     cache_dir = tmp_path / ".statue"
     with pytest.raises(
         InvalidConfiguration, match="^Got unexpected runner mode in configuration: BLA$"
     ):
-        ConfigurationBuilder.from_dict(
+        Configuration.from_dict(
             cache_dir=cache_dir, statue_config_dict={GENERAL: {MODE: "bla"}}
         )
